refactor: extract low-churn domain mixins from ICModelController (steps 1-2)
Move chemistry, sensor, cover, schedule, circuit-group, and light helpers into src/pyintellicenter/_mixins/ composed into ICModelController. Behavior-preserving: public API and __all__ unchanged, signatures identical. Heater/pump/body/system groups deferred to later steps.
